refactor functions for meal diary page
