+++ lib/event-listeners.js
@@ -1,13 +1,13 @@
-import {getMeals,
-        deleteCall,
-        deleteFoodFromMeals} from './fetch-crud'
-import findFoodFromMeals from './meals'
+import {findFoodsInMeals,
+        deleteFood,
+        deleteFoodFromMeals,
+          populateFoodFromMeals} from './fetch-meals'
 
 $('#food-list').on('click', '.delete-food-button', function (event) {
   event.preventDefault()
   event.target.parentElement.parentElement.remove()
-  getMeals(event.target.id)
-  deleteCall(event.target.id)
+  findFoodsInMeals(event.target.id)
+  deleteFood(event.target.id)
 })
 
 $('.meal-table').on('click', '.delete-meal-food-button', function (event) {
@@ -16,4 +16,9 @@ $('.meal-table').on('click', '.delete-meal-food-button', function (event) {
   let mealId = event.currentTarget.parentElement.id
   event.target.parentElement.parentElement.remove()
   deleteFoodFromMeals(mealId, foodId)
+})
+
+$("#new-food-btn").click(function(event) {
+  event.preventDefault()
+  window.location="./foods.html"
 })
