+++ lib/foods.js
@@ -26,13 +26,6 @@ const orderById = (a, b) => {
   return 0;
  }
 
-// $('#food-list').on('click', '.delete-food-button', function (event) {
-//   event.preventDefault()
-//   event.target.parentElement.parentElement.remove()
-//   getMeals(event.target.id)
-//   deleteCall(event.target.id)
-// })
-
 $('#food-list').on('click', '.food-name-field', function (event){
   let tableCell = $(`<td class="food-name-field" id='${event.currentTarget.id}'><input type="text" name="food-name" value="${event.target.textContent}"
   id="food-name-edit" required ></td>`)
@@ -58,36 +51,6 @@ $('#food-list').on('click', '.food-calorie-field', function (event){
     exitEvent.stopImmediatePropagation()
   });
 })
-
-// function deleteCall (id) {
-//   return fetch(`https://vast-retreat-17218.herokuapp.com/api/v1/foods/${id}`, {
-//     method: 'DELETE'
-//   }).catch(error => console.error({ error }))
-// }
-
-// const getMeals = foodId => {
-//   fetch('https://vast-retreat-17218.herokuapp.com/api/v1/meals')
-//     .then(response => response.json())
-//     .then(parsedResponse => {
-//       parsedResponse.forEach(function (element) {
-//         element.foods.forEach(function (food) {
-//           // this is an array
-//           if (food.id == foodId) {
-//             deleteFoodFromMeals(element.id, food.id)
-//           }
-//         })
-//       })
-//     })
-// }
-
-// const deleteFoodFromMeals = (mealID, foodId) => {
-//   fetch(
-//     `https://vast-retreat-17218.herokuapp.com/api/v1/meals/${mealID}/foods/${foodId}`,
-//     { method: 'DELETE' }
-//   ).catch(error => console.error({ error }))
-// }
-
-
 
 $("form").submit(function( event ) {
   event.preventDefault()
@@ -130,31 +93,6 @@ const postConfig = (foodInfo) => {
   }
 }
 
-
-
-const searchFunction = (event) => {
-
-  // let input = $('#food-filter').val()
-  // let listItems = $("tr")
-  // let category = document.querySelector('.mw-category');
-// let links = category.querySelectorAll('a')
-// let links = Array.from(document.querySelectorAll(".mw-category a"));
-// let de = links
-//               .map(link => link.textContent)
-//               .filter(streetName => streetName.includes('de'));
-
-
-  // listItems.forEach(function)
-  // for (i = 0; i < list.length; i++) {
-  //   p = listItems[i].getElementsByTagName("p")[0]
-  //   if (p.innerHTML.toUpperCase().indexOf(input > -1)) {
-  //     listItems[i].style.display = ""
-  //   } else {
-  //     listItems[i].style.display = "none"
-  //   }
-  //   }
-}
-
 $('#food-filter').on('keyup', function (event) {
   let input = event.currentTarget.value.toUpperCase()
   let i = input.length
@@ -168,7 +106,6 @@ $('#food-filter').on('keyup', function (event) {
     }
   })
 })
-
 
 const processResponse = (response) => {
   return response.json()
