+++ lib/html-sources.js
@@ -1,0 +1,21 @@
+// let totalCalCount
+// let remainingDailyCals
+//
+// const totalCaloriesHtml = (totalCalCount, remainingDailyCals) => {
+//   return `
+//   <tr id="goal-calories">
+//     <td><p>Goal Calories</p></td>
+//     <td><p>2000</p></td>
+//   </tr>
+//   <tr id="consumed-calories">
+//     <td><p>Consumed Calories</p></td>
+//     <td id="total-consumed-cals"><p>${totalCalCount}</p></td>
+//   </tr>
+//   <tr id="remaining-total-calories">
+//     <td><p>Remaining Calories</p></td>
+//     <td id="total-remaining-cals"><p>${remainingDailyCals}</p></td>
+//   </tr>
+//   `
+// }
+//
+// module.exports = totalCaloriesHtml
