+++ lib/index.js
@@ -1,3 +1,5 @@
 var foods = require("./foods")
 var meals = require('./meals')
 var listener = require('./event-listeners')
+var fetch = require('./fetch-meals')
+var functions = require('./shared-functions')
